fix: Android seekTo method uses long instead of int

+++ flutter_vlc_player/android/src/main/java/software/solid/fluttervlcplayer/FlutterVlcPlayer.java
@@ -348,7 +348,7 @@ final class FlutterVlcPlayer implements PlatformView {
         return mediaPlayer.getRate();
     }
 
-    void seekTo(int location) {
+    void seekTo(long location) {
         if (mediaPlayer == null) return;
 
         mediaPlayer.setTime(location);
